No log message.
git-svn-id: svn://orca.physics.unc.edu/Orca@343 fd86f90f-2c0a-422a-b972-84106d8db31b

+++ Source/Objects/Hardware/Camac/DGF4c/ORDGF4cModel.h
@@ -88,7 +88,7 @@ union {
 		unsigned long runBehaviorMask; //synchwait, insynch, etc.
 		short decimation;
 		long numOscPoints;
-		int oscModeData[4][8192];
+		unsigned short oscModeData[4][8192];
 		unsigned short oscEnabledMask;
 		float energyRiseTime[4];
 		float energyFlatTop[4];
